Update double to float in change_column

+++ db/migrate/20191025080018_add_extra_data_to_listings.rb
@@ -4,7 +4,7 @@ class AddExtraDataToListings < ActiveRecord::Migration[6.0]
     add_column :listings, :location, :integer
 
     # we have to make the columns before we can make them not null
-    change_column :listings, :price, :double, null: false
+    change_column :listings, :price, :float, null: false
     change_column :listings, :location, :integer, null: false
     
     add_column :listings, :amount, :integer, default: 1
